requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
tensorflow==2.16.1
numpy==1.26.4
python-multipart==0.0.9
jinja2==3.1.4
Pillow==10.4.0
icrawler==0.6.6
imagehash==4.3.1
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
jupyter>=1.0.0
ipykernel>=6.0.0
gradio>=4.0.0
